fix(runtime): distinguish unobserved provider cost
