+++ crates/harness-server/src/workflow_runtime_worker/turn_engine/mod.rs
@@ -5,6 +5,7 @@
 //! GH-1434 removes that path.
 
 pub(crate) mod helpers;
+mod runtime_usage;
 pub(crate) mod turn_lifecycle;
 
 #[cfg(test)]
