+++ crates/harness-workflow/src/runtime/tests/command_dispatcher_budget.rs
@@ -20,6 +20,7 @@ fn budget_usage_upsert(workflow_id: &str, cost_usd_micros: u64) -> RuntimeUsageU
         candidate_count: None,
         metrics: RuntimeUsageMetrics::default(),
         cost_usd_micros,
+        cost_usd_observed: true,
         reported_at: Utc::now(),
     }
 }
